Simplify utils
